Sync Maig-66 with Changes in Main

+++ lib/actions/delete-instruction.ts
@@ -1,18 +1,38 @@
 'use server'
+import {Instruction} from '@prisma/client'
 import {revalidatePath} from 'next/cache'
 import {redirect} from 'next/navigation'
 import prisma from '~/prisma'
+import {getCurrentUser} from '~/utils/session'
 
 export default async function deleteInstruction(
 	teamSlug: string,
-	projectId: string,
-	instructionId: string
+	instruction: Instruction
 ) {
-	await prisma.instruction.delete({
-		where: {
-			id: instructionId
+	const user = await getCurrentUser()
+	if (!user)
+		return {
+			message: 'Unauthorized, no session',
+			type: 'error'
 		}
-	})
-	revalidatePath(`/${teamSlug}/project/${projectId}/instructions`)
-	redirect(`/${teamSlug}/project/${projectId}/instructions`)
+
+	try {
+		await prisma.instruction.delete({
+			where: {
+				id: instruction.id
+			}
+		})
+		revalidatePath(`/${teamSlug}/project/${instruction.projectId}/instructions`)
+		redirect(`/${teamSlug}/project/${instruction.projectId}/instructions`)
+	} catch (err) {
+		if (err instanceof Error)
+			return {
+				message: err.message,
+				type: 'error'
+			}
+		return {
+			message: `Unexpected error: ${JSON.stringify(err)}`,
+			type: 'error'
+		}
+	}
 }
